+++ src/app/app.module.ts
@@ -1,17 +1,19 @@
+import { LoginPageComponent } from './components/login-page/login-page.component';
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { PokenewsComponent } from './components/pokenews/pokenews.component';
-import { RegisterComponent } from './register/register.component';
+import { RegisterComponent } from './components/register/register.component';
 import { HttpClientModule } from '@angular/common/http';
 
 @NgModule({
   declarations: [
     AppComponent,
     PokenewsComponent,
-    RegisterComponent
+    RegisterComponent,
+    LoginPageComponent
 
   ],
   imports: [
